Add RasperryPi-specific compilation option for Kodi
This commit adds `BR2_PACKAGE_RPI_USERLAND` to `rpi2` and `rpi3` configs
(it was already present in `rpi1` config.

It makes Buildroot build Kodi with the `CORE_SYSTEM_NAME=rbpi`
compilation option. See upstream `kodi.mk` for details:
https://github.com/buildroot/buildroot/blob/2018.02.3/package/kodi/kodi.mk

It should fix a segmentation fault when starting Kodi on RasPi 2 and 3,
when the collection contains at least 1 element (so, not reproducible
on a vanilla Recalbox installation: you need to add at least a TV show
or a movie and scrape it to make Kodi segfault):

```bash
$ LD_LIBRARY_PATH="/usr/lib/mysql" /usr/lib/kodi/kodi.bin --standalone -fs
Segmentation fault
```

> Where does the above command comes from?
> 1. EmulationStation starts `configgen -system 'kodi' -rom ''`:
>     https://gitlab.com/recalbox/recalbox-emulationstation/blob/46281da/es-app/src/recalbox/RecalboxSystem.cpp#L234
> 2. Configgen starts `kodilauncher.sh`:
>     https://gitlab.com/recalbox/recalbox-configgen/blob/b2cd1c6/configgen/recalboxFiles.py#L21
> 3. `kodilauncher.sh` starts Kodi with the above command:
>     https://gitlab.com/recalbox/recalbox/blob/069fb97/board/recalbox/fsoverlay/recalbox/scripts/kodilauncher.sh#L49